pyproject.toml
[project]
name = "llmops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

# langchain-core>0.3.0 breaks mlflow.langchain, for now
dependencies = [
    "faiss-cpu>=1.9.0",
    "langchain-community>=0.2.17",
    "langchain-core<0.3.0",
    "langchain-openai>=0.1.25",
    "langchain<0.3.5",
    "langgraph>=0.2.39",
    "mlflow[langchain]>=2.17.1",
    "openai>=1.52.2",
    "tiktoken>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
